breaking out anagram logic to new component
